Address PR #294 review feedback
Move `delete_app_container_profile` from `src/wxc/src/main.rs` into
`src/wxc_common/src/appcontainer_runner.rs`, next to `AppContainerScriptRunner`
(which owns the create/setup path). Both ends of the AppContainer profile
lifecycle now live in the same module.

While moving the function, replace the open-coded
`FileSystemBfsManager::new(...) + remove_configuration(...)` block with
`FileSystemBfsManager::clear_policy(name, logger)`. The original code created a
fresh manager and called `remove_configuration` without first setting
`configured = true`, so the early-return in `remove_configuration` made the
BFS-clear step a silent no-op for `wxc-exec --delete`. `clear_policy` is the
helper purpose-built for externally-created sandboxes and sets
`configured = true` itself, so the cleanup actually runs now.
